% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{altgreaterthannull}
\alias{altgreaterthannull}
\title{Calculates P-Value when the alternate hypothesis is greater than the null}
\usage{
altgreaterthannull(xbar, mu, sd, n)
}
\arguments{
\item{xbar}{Sample mean}

\item{mu}{Hypothesized sample mean}

\item{sd}{Standard Deviation}

\item{n}{Sample Size}
}
\value{

}
\description{
Calculates P-Value when the alternate hypothesis is greater than the null
}
