% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{altlessthannull}
\alias{altlessthannull}
\title{Computes P-value when the alternate hypothesis is less than the null}
\usage{
altlessthannull(xbar, mu, sd, n)
}
\arguments{
\item{xbar}{Sample mean}

\item{mu}{Hypothesized sample mean}

\item{sd}{Standard Deviation}

\item{n}{Sample Size}
}
\value{

}
\description{
Computes P-value when the alternate hypothesis is less than the null
}
