% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{altnotnull}
\alias{altnotnull}
\title{Computes P-value when the alternate hypothesis =/ the null}
\usage{
altnotnull(xbar, mu, sd, n)
}
\arguments{
\item{xbar}{Sample mean}

\item{mu}{Hypothesized sample mean}

\item{sd}{Standard Deviation}

\item{n}{Sample Size}
}
\value{

}
\description{
Computes P-value when the alternate hypothesis =/ the null
}
