% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceIntervals.R
\name{CI.diff}
\alias{CI.diff}
\title{Margin of error for difference between population means}
\usage{
CI.diff(CI, x1, x2)
}
\arguments{
\item{CI}{Confidence interval expressed as a decimal}

\item{x1}{Vector (bigger values)}

\item{x2}{Vector (smaller values)}
}
\value{

}
\description{
Margin of error for difference between population means
}
