% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceIntervals.R
\name{CI.paired.diff}
\alias{CI.paired.diff}
\title{Margin of Error for difference between paired population means}
\usage{
CI.paired.diff(CI, ssd1, ssd2, n1, n2, large = TRUE)
}
\arguments{
\item{CI}{Confidence interval expressed as a decimal}

\item{ssd1}{Sample Standard Deviation 1}

\item{ssd2}{Sample Standard Deviation 2}

\item{n1}{Sample size 1}

\item{n2}{Sample size 2}
}
\value{

}
\description{
Margin of Error for difference between paired population means
}
