% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceIntervals.R
\name{CI.prop}
\alias{CI.prop}
\title{Margin of error for a single large population proportion}
\usage{
CI.prop(CI, p1, n)
}
\arguments{
\item{CI}{Confidence interval expressed as a decimal}

\item{p1}{Number of subjects 1 (numerator of proportion)}

\item{n}{Sample size}
}
\value{

}
\description{
Margin of error for a single large population proportion
}
\examples{
To calculate bounds, subtract/add this functions output to p=(p1+2)/(n+4)
}
