% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{detsiglevel}
\alias{detsiglevel}
\title{Determines the significance level}
\usage{
detsiglevel(p, H0, n)
}
\arguments{
\item{p}{proportion as a decimal}

\item{H0}{null hypothesis threshold}

\item{n}{Sample size}
}
\value{

}
\description{
Determines the significance level
}
