% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceIntervals.R
\name{error.margin}
\alias{error.margin}
\title{Error margin for normal distribution CI with a population standard deviation where n is large}
\usage{
error.margin(CI, n, sd, large = TRUE)
}
\arguments{
\item{CI}{Confidence interval expressed as a decimal}

\item{n}{Sample size}

\item{sd}{Population standard deviation}
}
\value{

}
\description{
Error margin for normal distribution CI with a population standard deviation where n is large
}
