% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTesting.R
\name{H.diff.prop.nopopvariance}
\alias{H.diff.prop.nopopvariance}
\title{Calculates Z test stat for a difference between two (paired) population proportions, population variances unknown and not assumed equal. Gives inputs for pnorm}
\usage{
H.diff.prop.nopopvariance(xbar1, xbar2, ssd1, ssd2, n1, n2, islarge = TRUE)
}
\arguments{
\item{xbar1}{Average for bigger data}

\item{xbar2}{Average for smaller data}

\item{ssd1}{Sample Standard Deviation of the bigger data}

\item{ssd2}{Sample Standard Deviation of the smaller data}

\item{n1}{Bigger sample size}

\item{n2}{Smaller sample size}
}
\value{

}
\description{
Calculates Z test stat for a difference between two (paired) population proportions, population variances unknown and not assumed equal. Gives inputs for pnorm
}
