% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTesting.R
\name{H.diff.small.nopopvariance}
\alias{H.diff.small.nopopvariance}
\title{Hypothesis test for a VECTOR that is population mean difference. Outputs values needed to put into pt}
\usage{
H.diff.small.nopopvariance(x, mu)
}
\arguments{
\item{x}{Vector}

\item{mu}{Hypothesized sample mean}
}
\value{

}
\description{
Hypothesis test for a VECTOR that is population mean difference. Outputs values needed to put into pt
}
