% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTesting.R
\name{H.paired.prop.l.nv}
\alias{H.paired.prop.l.nv}
\title{Calculates Z test statistic for a difference between two (paired) popultation proportions}
\usage{
H.paired.prop.l.nv(p1, s1, p2, s2)
}
\arguments{
\item{p1}{How many out of the sample size for first data}

\item{s1}{Sample size for first data}

\item{p2}{How many out of the sample size for first data}

\item{s2}{Sample size for second data}
}
\value{

}
\description{
Calculates Z test statistic for a difference between two (paired) popultation proportions
}
