% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTesting.R
\name{H.prop.l.nv}
\alias{H.prop.l.nv}
\title{Calculates Z test statistic for a population proportion, variance is unknown, N is large}
\usage{
H.prop.l.nv(p1, n, p0)
}
\arguments{
\item{p1}{The numerator of the actual proportion}

\item{n}{The sample size (denominator of the actual proportion)}

\item{p0}{The wanted proportion as a decimal}
}
\value{

}
\description{
Calculates Z test statistic for a population proportion, variance is unknown, N is large
}
