% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTesting.R
\name{H.Vectorsgiven.paired.small.nopopvariance}
\alias{H.Vectorsgiven.paired.small.nopopvariance}
\title{Calculates t test for paired small vectors}
\usage{
H.Vectorsgiven.paired.small.nopopvariance(x1, x2)
}
\arguments{
\item{x1}{First Vector}

\item{x2}{Second Vector}
}
\value{

}
\description{
Calculates t test for paired small vectors
}
