% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{heart}
\alias{heart}
\title{Calculates basic test statistic, variance is known, n is large}
\usage{
heart(xbar, mu, sd, n)
}
\arguments{
\item{xbar}{Sample mean}

\item{mu}{Hypothesized sample mean}

\item{sd}{Standard deviation}

\item{n}{Sample size}
}
\value{
The test statistic
}
\description{
Calculates basic test statistic, variance is known, n is large
}
