% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{lunknown}
\alias{lunknown}
\title{Calculates P value when the alternate hypothesis is less than null, sample size is small, population variance is unknown}
\usage{
lunknown(x, mu)
}
\arguments{
\item{x}{Vector}

\item{mu}{Hypothesized sample mean}
}
\value{

}
\description{
Calculates P value when the alternate hypothesis is less than null, sample size is small, population variance is unknown
}
