% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{nunknown}
\alias{nunknown}
\title{Calculates P value when the alternate hypothesis =/ null, sample size is small, population variance is unknown}
\usage{
nunknown(x, mu)
}
\arguments{
\item{x}{Vector}

\item{mu}{Hypothesized sample mean}
}
\value{

}
\description{
Calculates P value when the alternate hypothesis =/ null, sample size is small, population variance is unknown
}
