% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{paireddatasmallsample}
\alias{paireddatasmallsample}
\title{Calculates t test for paired vectors}
\usage{
paireddatasmallsample(x1, x2)
}
\arguments{
\item{x1}{First Vector}

\item{x2}{Second Vector}
}
\value{

}
\description{
Calculates t test for paired vectors
}
