% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{phatsiglevel}
\alias{phatsiglevel}
\title{Calculates the value of p hat for calculating significance level}
\usage{
phatsiglevel(H0, n)
}
\arguments{
\item{H0}{null hypothesis threshold}

\item{n}{sample size}
}
\value{

}
\description{
Calculates the value of p hat for calculating significance level
}
