% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{ptstatnissmall}
\alias{ptstatnissmall}
\title{Calculates test statistic for paired data, variance is unknown, n is large}
\usage{
ptstatnissmall(p1, s1, p2, s2)
}
\arguments{
\item{p1}{How many out of the sample size for first data}

\item{s1}{Sample size for first data}

\item{p2}{How many out of the sample size for first data}

\item{s2}{Sample size for second data}
}
\value{

}
\description{
Calculates test statistic for paired data, variance is unknown, n is large
}
