% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{ptstatvarunequal}
\alias{ptstatvarunequal}
\title{Calculates test statistic for paired data, variances unknown and not assumed equal}
\usage{
ptstatvarunequal(mu1, mu2, sd1, sd2, n1, n2)
}
\arguments{
\item{mu1}{Average for first data}

\item{mu2}{Average for second data}

\item{sd1}{Standard Deviation of the first data}

\item{sd2}{Standard Deviation of the second data}

\item{n1}{First sample size}

\item{n2}{Second sample size}
}
\value{

}
\description{
Calculates test statistic for paired data, variances unknown and not assumed equal
}
