% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{sdphatsiglvl}
\alias{sdphatsiglvl}
\title{Standard deviation p hat for significance level}
\usage{
sdphatsiglvl(p, n)
}
\arguments{
\item{p}{proportion to be tested, a decimal}

\item{n}{Sample size}
}
\value{

}
\description{
Standard deviation p hat for significance level
}
