% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTesting.R
\name{test.stat}
\alias{test.stat}
\title{Calculates basic test statistic Z or T}
\usage{
test.stat(xbar, mu, sd, n, large = TRUE, comparison = "notequal")
}
\arguments{
\item{xbar}{Sample mean}

\item{mu}{Hypothesized sample mean}

\item{sd}{Standard deviation}

\item{n}{Sample size}

\item{large}{Treat the sample as large?}

\item{comparison}{notequal, greaterthan or lessthan}
}
\value{
The test statistic
}
\description{
Calculates basic test statistic Z or T
}
