% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{tstatnislarge}
\alias{tstatnislarge}
\title{Calculates test statistic, variance is unknown, N is large}
\usage{
tstatnislarge(p1, n, p0)
}
\arguments{
\item{p1}{The numerator of the actual proportion}

\item{n}{The sample size (denominator of the actual proportion)}

\item{p0}{The wanted proportion as a decimal}
}
\value{

}
\description{
Calculates test statistic, variance is unknown, N is large
}
