% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{unknownpopvarsmallsamplesizetvalue}
\alias{unknownpopvarsmallsamplesizetvalue}
\title{T for computing P value of a vector with a hypothesized mean}
\usage{
unknownpopvarsmallsamplesizetvalue(x, mu)
}
\arguments{
\item{x}{Vector}

\item{mu}{Hypothesizedsample mean}
}
\value{

}
\description{
T for computing P value of a vector with a hypothesized mean
}
