% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTesting.R
\name{z.ignore.calculates.p.hat.for.detsiglevel}
\alias{z.ignore.calculates.p.hat.for.detsiglevel}
\title{Calculates the value of p hat for calculating significance level}
\usage{
z.ignore.calculates.p.hat.for.detsiglevel(H0, n)
}
\arguments{
\item{H0}{null hypothesis threshold}

\item{n}{sample size}
}
\value{

}
\description{
Calculates the value of p hat for calculating significance level
}
