% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothesisTesting.R
\name{z.ignore.calculates.p.hat.sd.for.detsiglevel}
\alias{z.ignore.calculates.p.hat.sd.for.detsiglevel}
\title{Standard deviation p hat for significance level}
\usage{
z.ignore.calculates.p.hat.sd.for.detsiglevel(p, n)
}
\arguments{
\item{p}{proportion to be tested, a decimal}

\item{n}{Sample size}
}
\value{

}
\description{
Standard deviation p hat for significance level
}
