% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{heart}
\alias{heart}
\title{Calculates test statistic, N is large, standard deviation is known}
\usage{
heart(xbar, mu, sd, n)
}
\arguments{
\item{xbar}{Sample mean}

\item{mu}{Population mean}

\item{sd}{Standard deviation}

\item{n}{Observation count}
}
\value{
The test statistic
}
\description{
Calculates test statistic, N is large, standard deviation is known
}
