% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{unknownpopvarsmallsamplesizetvalue}
\alias{unknownpopvarsmallsamplesizetvalue}
\title{Title}
\usage{
unknownpopvarsmallsamplesizetvalue(x, mu)
}
\arguments{
\item{mu}{}
}
\value{

}
\description{
Title
}
